src/requirements.txt
Django >=1.11, <2
django-admin-sortable2
django-allauth
django-anymail[mailgun]
django-ckeditor
django-cleanup
django-colorful
django-countries
django-crispy-forms
django-easy-select2==1.3.4
django-filter==1.1.0
django-flat-responsive
django-geoposition
django-image-cropping
django-import-export
django-modeltranslation
django-phonenumber-field
django-unused-media
django-widget-tweaks
django-wkhtmltopdf
djangorestframework
easy-thumbnails
mailchimp
Pillow
MySQL-python
